Add files via upload
Modify window is not complete yet, but the other things are fine. Make sure you have all the necessary programs to run the axidraw (everything listed on the top of the Spotexp_GUI. You need to download the modules that come with the inkscape app, which is made to run the axidraw, and the axidraw python API in order to run this properly.